+++ apps/escrow/src/hocs/index.tsx
@@ -8,6 +8,7 @@ import { Alert, alertStyles } from '@gear-js/ui';
 import { ComponentType } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { ADDRESS } from 'consts';
+import { WasmProvider } from 'context';
 
 function ApiProvider({ children }: ProviderProps) {
   return <GearApiProvider providerAddress={ADDRESS.NODE}>{children}</GearApiProvider>;
@@ -21,7 +22,7 @@ function AlertProvider({ children }: ProviderProps) {
   );
 }
 
-const providers = [BrowserRouter, AlertProvider, ApiProvider, AccountProvider];
+const providers = [BrowserRouter, AlertProvider, ApiProvider, AccountProvider, WasmProvider];
 
 function withProviders(Component: ComponentType) {
   return () => providers.reduceRight((children, Provider) => <Provider>{children}</Provider>, <Component />);