+++ apps/nft/src/assets/wasm/index.ts
@@ -1,3 +1,0 @@
-import nftMetaWasm from './nft.meta.wasm';
-
-export { nftMetaWasm };