+++ apps/social/src/components/layout/header/Header.tsx
@@ -1,19 +1,20 @@
-import { useApi } from '@gear-js/react-hooks';
 import { Logo } from './logo';
 import { Menu } from './menu';
 import { Account } from './account';
 import styles from './Header.module.scss';
 
-function Header() {
-  const { isApiReady } = useApi();
+type Props = {
+  isAccountVisible: boolean;
+};
 
+function Header({ isAccountVisible }: Props) {
   return (
     <header className={styles.header}>
       <nav className={styles.nav}>
         <Logo />
         <Menu />
       </nav>
-      {isApiReady && <Account />}
+      {isAccountVisible && <Account />}
     </header>
   );
 }