+++ apps/supply-chain/src/consts.ts
@@ -10,6 +10,7 @@ import { ReactComponent as infoSVG } from 'assets/images/actions/info.svg';
 
 const ADDRESS = {
   NODE: process.env.REACT_APP_NODE_ADDRESS as string,
+  DAPPS_API: process.env.REACT_APP_DAPPS_API_ADDRESS as string,
 };
 
 const LOCAL_STORAGE = {