+++ utils/gear-hooks/src/hooks/api/index.ts
@@ -1,4 +1,3 @@
-import { useBalanceSubscription } from './useBalanceSubscription';
 import { useMetadata } from './useMetadata';
 import { useReadState } from './useReadState';
 import { useSendMessage } from './useSendMessage';
@@ -11,7 +10,6 @@ import {
 } from './useCalculateGas';
 
 export {
-  useBalanceSubscription,
   useMetadata,
   useReadState,
   useSendMessage,